refactor: separate round class
