+++ src/main/java/racing/GrandPrix.java
@@ -25,15 +25,4 @@ public class GrandPrix {
         currentRound++;
     }
 
-    static class Round {
-        private boolean finished;
-
-        void play() {
-            finished = true;
-        }
-
-        boolean isFinished() {
-            return finished;
-        }
-    }
 }
