+++ src/main/java/racing/Round.java
@@ -1,0 +1,14 @@
+package racing;
+
+public class Round {
+    private boolean finished;
+
+    void play() {
+        finished = true;
+    }
+
+    boolean isFinished() {
+        return finished;
+    }
+
+}
